+++ certbot-dns-linode/certbot_dns_linode/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_linode.dns_linode` plugin.
