+++ certbot-dns-linode/docs/api/dns_linode.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_linode.dns_linode`
+:mod:`certbot_dns_linode._internal.dns_linode`
 ------------------------------------------------
 
-.. automodule:: certbot_dns_linode.dns_linode
+.. automodule:: certbot_dns_linode._internal.dns_linode
    :members:
